[Misc] Fix Sonar Cast one of the operands of this addition operation to a "long"

+++ xwiki-platform-core/xwiki-platform-tree/xwiki-platform-tree-api/src/main/java/org/xwiki/tree/AbstractTreeNode.java
@@ -147,7 +147,7 @@ public abstract class AbstractTreeNode implements TreeNode
         }
 
         int start = Math.clamp(offset, 0, list.size());
-        int end = Math.clamp(start + limit, start, list.size());
+        int end = Math.clamp((long) start + limit, start, list.size());
         return list.subList(start, end);
     }
